Modified the OAuth2Request to be built from the ZF2 Request.
There was an issue present where the OAuth2Request was reading php://input
and when ZF2 tried to read php://input it was null because php://input can
only be read once per request.

I modified the AuthorizationListner to create the OAuth2Request manually and
set the content from the ZF2 Request.