Implemented authentication map
The authentication map is a map of API/module => authentication type.
This allows per-API authentication.

However, in order to keep this backwards compatible with pre-1.1, the
following scenarios needed to pass:

- if
  - no authentication schemes are defined
  - do not perform authentication
- if
  - a single authentication scheme is defined
  - and no map is present
  - use the defined authentication
- if
  - multiple authentication schemes are defined
  - and no map is present
  - do not perform authentication
- if
  - one or more authentication schemes are defined
  - and a map is present
  - and the current API/module does not have an entry in the map
  - do not perform authentication
- if
  - one or more authentication schemes are defined
  - and a map is present
  - and the current API/module has an entry in the map
  - that does not match one of the authentication schemes
  - do not perform authentication
- if
  - one or more authentication schemes are defined
  - and a map is present
  - and the current API/module has an entry in the map
  - that matches one of the authentication schemes
  - use the defined authentication

This will keep current applications working, while simultaneously allowing them
to start adopting the new mapping functionality. The admin API will soon be
updated to write in this format, and will adapt an existing application to
setup an authentication map for each defined API and version.