Adapt #46 for develop branch
The develop branch introduces Authentication Adapters, which necessitates a
change in the testing strategy. Additionally, it makes sense to return earlier
in the method if the service exists.

A number of tests were incorrectly using `sqlite:memory:` instead of
`sqlite::memory:` as a PDO DSN; fixed to ensure tests continue to work.