Cache results of the NamedOAuth2ServerFactory
One root cause of #70 is that the factories returned by the various
`ZF\OAuth2\Service\OAuth2Server` implementations were not caching the
`OAuth2\Server` instances they generated. This patch ensures that the
zf-mvc-auth implementation does, and is a complement to
zfcampus/zf-oauth2#95.

Additionally, in order to test this, I had to implement a fix for #69:
fixing `ZF\MvcAuth\Factory\OAuth2ServerFactory` to allow named storage
services or an array of named storage services.