Change priorities of authentication events
zf-versioning uses priorities -40 and -41 in order to rewrite the matched
controller based on versioning information. Since per-API authentication is
based on the combination of the API namespace **and** version, authentication
listeners _must_ trigger _after_ versioning. As such, this patch moves the
priorities for the authentication and authentication.post events to -50 and -51,
respectively.