[#83] Return a 401 for invalid OAuth2 credentials
`OAuth2\Server` returns boolean false for each of:

- no token present
- invalid and/or expired token present
- token invalid for current scope

The first case, is expected; it's what will happen in a public API when no
credentials are provided, and should result in marshaling a `GuestIdentity`.

The second two cases, however, should result in returning the appropriate status
code and headers. For this to happen, we must introspect the response composed
in the server:

- First, to see if we have a 401 or 403 status, and
- second, to see if an error has been reported to the response.

If both conditions occur, we merge the `OAuth2\Response` status and headers to
the MVC response and return it immediately.

Otherwise, we only merge any headers present (typically, the `WWW-Authenticate`
header, which will be a prompt for clients to know that authentication is
possible), and return a `GuestIdentity`.