Only register listeners for HTTP requests
Based on a report from @jguittard via ContinuousPHP, registering the
mvc-auth listeners can have ill effects if not in an HTTP environment and/or the
local configuration for OAuth2 is missing (which is often the case when
packaging for deployment). This patch adds code to `onBootstrap()` to return
early, without registering listeners, if a non-HTTP request is made.