Ensure ACL rules work when dash is used for namespace separator in definitions
The Apigility Admin UI/API currently uses dash separators when defining ACL
rules. This can lead to issues with the ACLs generated, as zend-mvc always uses
namespace separators. This change ensures that the rules are correctly
populated in the ACL during creation.

Another patch will be issued against zf-apigility-admin to ensure namespace
separators are used in the future; in the meantime, this patch ensures
zf-mvc-auth will work correctly.